requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
Pillow>=10.4.0
numpy>=1.26.4
opencv-python>=4.10.0.84
PyYAML>=6.0.2
torch>=2.3.0
